FEATURE(redis): Manage redis server `slave-priority`
 ##### SUMMARY

This change allows to:
  - define a instance slave only
  - prior a pool of master (typically in a particular location: DC, region, ...)

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION
https://redis.io/topics/sentinel

```
Redis instances have a configuration parameter called slave-priority. This information is exposed by Redis slave instances in their INFO output, and Sentinel uses it in order to pick a slave among the ones that can be used in order to failover a master:

* If the slave priority is set to 0, the slave is never promoted to master.
* Slaves with a lower priority number are preferred by Sentinel.

For example if there is a slave S1 in the same data center of the current master, and another slave S2 in another data center, it is possible to set S1 with a priority of 10 and S2 with a priority of 100, so that if the master fails and both S1 and S2 are available, S1 will be preferred.

For more information about the way slaves are selected, please check the slave selection and priority section of this documentation.
```